Remove obsolete patching from BaseTransientModelFormSet._construct_form
Ever since Django 1.7 it's been unnecessary to special-case the model
instance fetching code to retrieve from the queryset instead of the database,
because Django does this itself (via the _existing_object method which builds
a lookup table of instances from the queryset). The only thing we need
BaseTransientModelFormSet to do now is catch `pk == ''` as a special case,
because that will break to_python.