Add Django to setup.py
The current setup.py doesn't include Django in its install_requires.
This makes it a little harder to do local development on this
repository alone, because you have to manually install Django.
Additionally, this would make it easier for developers to quickly see
which versions of Django are supported by django-modelcluster.

With this change I've also removed pytz from setup.py and tox.ini, as it
is automatically installed by Django.