Avoid accessing live queryset on unsaved instances
In cases where a DeferringRelatedManager was read on an in-memory instance before being written to, it was accessing the real manager on an unsaved instance. This was tolerated up to Django 4.0, but it's a hard fail as of https://github.com/django/django/commit/7ba6ebe9149ae38257d70100e8bfbfd0da189862.